+++ CHANGELOG.md
@@ -13,6 +13,8 @@ This project uses [Semantic Versioning][sv].
   where each integer is a bitfield encoding the passage directions.
 - Add `Bitmask#deserialize` to decode an array of bitmasked integers into a
   maze.
+- `BinaryTree#carve` accepts an argument to set the seed for the random number
+  generator
 
 ### Changed
 - Update Usage in README with namespace changes.